Create Voltage init controller and service
